chore: renamed ./edit-panel-components to ./edit-panel.
