+++ src/components/resume-builder-v2/components/edit-panel/TitleEditBox.tsx
